requirements.txt
flask==2.0.1
spacy==3.5.3
scikit-learn==1.3.0
python-docx==0.8.11
pdfminer.six==20221105
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
python-multipart==0.0.6